server/requirements.txt
django>=1.5
djangorestframework
babel
South

factory_boy

#django-oauth2-provider

#requests-oauth2
